+++ src/data.h
@@ -6,15 +6,15 @@
 
 struct Flower {
     std::string specie;
-    float sepal_length;
-    float sepal_width;
-    float petal_length;
-    float petal_width;
+    double sepal_length;
+    double sepal_width;
+    double petal_length;
+    double petal_width;
 };
 
 std::vector<Flower> data(std::string filePath);
 
-std::vector<std::vector<float>> prepareTrainingData(const std::vector<Flower>& flowers);
+std::vector<std::vector<double>> prepareTrainingData(const std::vector<Flower>& flowers);
 
 std::string trim(const std::string& s);
