+++ src/network/multilayerPerceptron.cpp
@@ -5,15 +5,15 @@
 
 MultilayerPerceptron::MultilayerPerceptron(
     std::vector<int> layerSizes,
-    std::function<float(float)> activationFunction
+    std::function<double(double)> activationFunction
 )
     : Network(layerSizes, activationFunction)
 {
     id = global_id_counter++;
 }
 
-int MultilayerPerceptron::softMaxInPlace(std::vector<float>& output) {
-    float sumExp = 0.0f;
+int MultilayerPerceptron::softMaxInPlace(std::vector<double>& output) {
+    double sumExp = 0.0f;
 
     for (size_t i = 0; i < output.size(); ++i) {
         output[i] = std::exp(output[i]);
@@ -27,8 +27,8 @@ int MultilayerPerceptron::softMaxInPlace(std::vector<float>& output) {
     return 0;
 }
 
-const std::vector<float> MultilayerPerceptron::forward(const std::vector<float>& inputs) {
-    std::vector<float> bufferA, bufferB;
+const std::vector<double> MultilayerPerceptron::forward(const std::vector<double>& inputs) {
+    std::vector<double> bufferA, bufferB;
     bufferA.resize(maxLayerSizes);
     bufferB.resize(maxLayerSizes);
     inputLayer.forward(inputs, bufferA);
@@ -43,22 +43,22 @@ const std::vector<float> MultilayerPerceptron::forward(const std::vector<float>&
     return bufferA;
 }
 
-float MultilayerPerceptron::computeLoss(const std::vector<float>& predicted, int target) {
+double MultilayerPerceptron::computeLoss(const std::vector<double>& predicted, int target) {
     // categorical cross-entropy loss
-    float p = predicted[target];
+    double p = predicted[target];
 
     // Éviter log(0)
-    float loss = -std::log(p + 1e-15f);
+    double loss = -std::log(p + 1e-15f);
 
     return loss;
 }
 
-void MultilayerPerceptron::training(std::vector<std::vector<float>> trainingData, int epochs, float learningRate) {
+void MultilayerPerceptron::training(std::vector<std::vector<double>> trainingData, int epochs, double learningRate) {
 
     std::random_device rd;
     std::mt19937 g(rd());
 
-    std::vector<float> inputsTemp;
+    std::vector<double> inputsTemp;
     inputsTemp.resize(maxLayerSizes);
     printf("%d", inputsTemp.size());
 
@@ -67,16 +67,16 @@ void MultilayerPerceptron::training(std::vector<std::vector<float>> trainingData
         //shuffle training data for each epoch
         std::shuffle(trainingData.begin(), trainingData.end(), g);
 
-        float epochLoss = 0.0f;
+        double epochLoss = 0.0f;
 
         for (const auto& dataPoint : trainingData) {
             // Assuming the last element is the target output
-            std::vector<float> inputs(dataPoint.begin(), dataPoint.end() - 1);
+            std::vector<double> inputs(dataPoint.begin(), dataPoint.end() - 1);
             int target = static_cast<int>(dataPoint.back());
 
-            const std::vector<float> output = forward(inputs);
+            const std::vector<double> output = forward(inputs);
 
-            float loss = computeLoss(output, target);
+            double loss = computeLoss(output, target);
             // printf("Output: %f\n", output[0]);
             epochLoss += loss;
 
@@ -86,11 +86,11 @@ void MultilayerPerceptron::training(std::vector<std::vector<float>> trainingData
 
                 if (nextLayer) {
                     for (size_t j = 0; j < layer->neuronLayer.size(); ++j) {
-                        float error = 0.0f;
+                        double error = 0.0f;
                         for (Neuron* nextNeuron : nextLayer->neuronLayer) {
                             error += nextNeuron-> weights[j] * nextNeuron->score;
                         }
-                        float derivative = layer->neuronLayer[j]->output * (1 - layer->neuronLayer[j]->output); // Sigmoid derivative
+                        double derivative = layer->neuronLayer[j]->output * (1 - layer->neuronLayer[j]->output); // Sigmoid derivative
                         layer->neuronLayer[j]->score = error * derivative;
                     }
                 }
@@ -107,7 +107,7 @@ void MultilayerPerceptron::training(std::vector<std::vector<float>> trainingData
                 }
 
                 for (Neuron* neuron : layer->neuronLayer) {
-                    // float gradient = neuron->score * neuron->output * (1 - neuron->output); // Sigmoid derivative
+                    // double gradient = neuron->score * neuron->output * (1 - neuron->output); // Sigmoid derivative
                     for (size_t w = 0; w < neuron->weights.size(); ++w) {
                         neuron->weights[w] -= learningRate * neuron->score * inputsTemp[w]; // Update weights
                     }
